Update library and restructure of component
