--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/.eslintrc.json
@@ -1,30 +1,0 @@
-{
-    "env": {
-      "browser": true,
-      "es2020": true
-    },
-    "extends": [
-      "eslint:recommended",
-      "plugin:@typescript-eslint/recommended"
-    ],
-    "globals": {
-      "ComponentFramework": true
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-      "ecmaVersion": 12,
-      "sourceType": "module"
-    },
-    "plugins": [
-      "@microsoft/power-apps",
-      "@typescript-eslint"
-    ],
-    "rules": {
-      "@typescript-eslint/no-unused-vars": "off",
-      "@typescript-eslint/no-explicit-any": ["off"],
-      "no-unused-vars": "off",
-     "@typescript-eslint/no-this-alias":"off",
-     "no-case-declarations":"off",
-     "prefer-const":"off"
-    }
-}
