+++ OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/.gitignore
@@ -1,14 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# dependencies
-/node_modules
-
-# generated directory
-**/generated
-
-# output directory
-/out
-
 # msbuild output directories
 /bin
 /obj
