+++ OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/ControlManifest.Input.xml
@@ -1,53 +1,14 @@
-<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <manifest>
-  <control namespace="ECC.PCF" constructor="DatePicker" version="0.0.1" display-name-key="DatePicker" description-key="DatePicker description" control-type="standard" >
-    <!--external-service-usage node declares whether this 3rd party PCF control is using external service or not, if yes, this control will be considered as premium and please also add the external domain it is using.
-    If it is not using any external service, please set the enabled="false" and DO NOT add any domain below. The "enabled" will be false by default.
-    Example1:
-      <external-service-usage enabled="true">
-        <domain>www.Microsoft.com</domain>
-      </external-service-usage>
-    Example2:
-      <external-service-usage enabled="false">
-      </external-service-usage>
-    -->
-    <external-service-usage enabled="false">
-      <!--UNCOMMENT TO ADD EXTERNAL DOMAINS
-      <domain></domain>
-      <domain></domain>
-      -->
-    </external-service-usage>
-    <!-- property node identifies a specific, configurable piece of data that the control expects from CDS -->
-    <property name="DatePicker" display-name-key="DatePicker_Key" description-key="DatePicker_Desc_Key" of-type="DateAndTime.DateOnly" usage="bound" required="true" />
-    <!--
-      Property node's of-type attribute can be of-type-group attribute.
-      Example:
-      <type-group name="numbers">
-        <type>Whole.None</type>
-        <type>Currency</type>
-        <type>FP</type>
-        <type>Decimal</type>
-      </type-group>
-      <property name="sampleProperty" display-name-key="Property_Display_Key" description-key="Property_Desc_Key" of-type-group="numbers" usage="bound" required="true" />
-    -->
+  <control namespace="DatePickerNamespace" constructor="DatePicker" version="0.0.2" display-name-key="DatePicker" description-key="DatePicker description" control-type="standard">
+    <property name="Field" display-name-key="Field" description-key="field_Desc_Key" of-type="SingleLine.Text" usage="bound" required="true" />
+    <property name="DateFormat" display-name-key="Date Format" description-key="Please select date format" of-type="Enum" usage="input" required="true" default-value="0">
+      <value name="YYYY-MM" display-name-key="YYYY-MM" description-key="YYYY-MM">0</value>
+      <value name="MM/DD" display-name-key="MM/DD" description-key="MM/DD">1</value>
+    </property>
     <resources>
-      <code path="index.ts" order="1"/>
-      <!-- UNCOMMENT TO ADD MORE RESOURCES
-      <css path="css/DatePicker.css" order="1" />
-      <resx path="strings/DatePicker.1033.resx" version="1.0.0" />
-      -->
+      <code path="index.ts" order="1" />
+      <css path="css/datepicker.css" order="1" />
     </resources>
-    <!-- UNCOMMENT TO ENABLE THE SPECIFIED API
-    <feature-usage>
-      <uses-feature name="Device.captureAudio" required="true" />
-      <uses-feature name="Device.captureImage" required="true" />
-      <uses-feature name="Device.captureVideo" required="true" />
-      <uses-feature name="Device.getBarcodeValue" required="true" />
-      <uses-feature name="Device.getCurrentPosition" required="true" />
-      <uses-feature name="Device.pickFile" required="true" />
-      <uses-feature name="Utility" required="true" />
-      <uses-feature name="WebAPI" required="true" />
-    </feature-usage>
-    -->
   </control>
 </manifest>
