--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/DatePicker/Solution/.gitignore
@@ -1,8 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# msbuild output directories
-/bin
-/obj
-
-# MSBuild Binary and Structured Log
-*.binlog
