--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/DatePicker/Solution/src/Other/Relationships.xml
@@ -1,2 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<EntityRelationships xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" />
