--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/DatePicker/css/datepicker.css
@@ -1,24 +1,0 @@
-.dnl-first-select{
-    width: 30%;
-    border: none;
-    appearance: none;
-    height: 35px;
-    padding: 5px;
-    font-weight: 600;
-    font-family: "Segoe UI";
-}
-
-.dnl-second-select{
-    width: 40%;
-    border:none;
-    appearance: none;
-    height: 35px;
-    padding: 5px;
-    font-weight: 600;
-    font-family: "Segoe UI";
-}
-
-.dnl-first-select:hover, .dnl-second-select:hover{
-    border: 1px solid rgb(133, 133, 133);
-    appearance: auto;
-}
