--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/DatePicker/src/Other/Customizations.xml
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ImportExportXml xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <Entities />
-  <Roles />
-  <Workflows />
-  <FieldSecurityProfiles />
-  <Templates />
-  <EntityMaps />
-  <EntityRelationships />
-  <OrganizationSettings />
-  <optionsets />
-  <CustomControls />
-  <SolutionPluginAssemblies />
-  <EntityDataProviders />
-  <Languages>
-    <Language>1033</Language>
-  </Languages>
-</ImportExportXml>
