--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "pcf-project",
-  "version": "1.0.0",
-  "description": "Project containing your PowerApps Component Framework (PCF) control.",
-  "scripts": {
-    "build": "pcf-scripts build",
-    "clean": "pcf-scripts clean",
-    "lint": "pcf-scripts lint",
-    "lint:fix": "pcf-scripts lint fix",
-    "rebuild": "pcf-scripts rebuild",
-    "start": "pcf-scripts start",
-    "start:watch": "pcf-scripts start watch",
-    "refreshTypes": "pcf-scripts refreshTypes"
-  },
-  "dependencies": {
-  },
-  "devDependencies": {
-    "@microsoft/eslint-plugin-power-apps": "^0.2.33",
-    "@types/node": "^18.19.31",
-    "@types/powerapps-component-framework": "^1.3.11",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
-    "eslint": "^8.57.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.1.1",
-    "pcf-scripts": "^1",
-    "pcf-start": "^1",
-    "typescript": "^4.9.5"
-  }
-}
