--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/pcfconfig.json
@@ -1,3 +1,0 @@
-{
-    "outDir": "./out/controls"
-}
