--- OFM.Infrastructure.PCF/Controls/DatePicker/DatePicker/tsconfig.json
@@ -1,6 +1,0 @@
-{
-    "extends": "./node_modules/pcf-scripts/tsconfig_base.json",
-    "compilerOptions": {
-        "typeRoots": ["node_modules/@types"]
-    }
-}
